feat: export types.d.ts
