+++ packages/theme-script/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hiogawa/theme-script",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "main": "./dist/index.js",
   "module": "./dist/index.js",
@@ -10,6 +10,9 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
+    },
+    "./types": {
+      "types": "./types.d.ts"
     },
     "./dist/main": {
       "import": "./dist/main.js",
@@ -22,7 +25,8 @@
     }
   },
   "files": [
-    "dist"
+    "dist",
+    "types.d.ts"
   ],
   "license": "MIT",
   "repository": {
