+++ packages/theme-script/types.d.ts
@@ -1,0 +1,3 @@
+// TODO: type-check with src/main.ts?
+declare const __themeGet: () => string;
+declare const __themeSet: (theme: string) => void;
